os-agent/requirements.txt
# LLM & Agent
anthropic>=0.25.0
langchain>=0.2.0
langchain-anthropic>=0.1.15
langchain-community>=0.2.0

# SSH 远程执行
paramiko>=3.4.0

# 系统监控（本地模式）
psutil>=5.9.0

# CLI 界面
rich>=13.7.0
click>=8.1.0

# Web 界面（可选）
fastapi>=0.111.0
uvicorn>=0.29.0
websockets>=12.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.0.0
